Fix bug that causes double counting when a running job is scraped more than once.
The stored version of the job was not updated, so the original counts were used
each time when comparing. This caused, for example, the number of processed datums
for a running job to increase every time a scrape happened, even if the counter
in the job had not gone up at all.